Finmish styled refactor
