--- src/components/main/Main.css
@@ -1,4 +1,0 @@
-.main {
-  display: flex;
-  flex-direction: column;
-}
